optionally preserve the order or ROIs in a zip
Preserving the order of ROIs in a ZIP is important for my application, but returning ROIs as a dictionary loses this order. The zip file list contains the dictionary keys as file names in the preserved order. Since each list item in the returned list is a dict with 'name' key, no data is lost by returning a list over a dict. Making it optional preserves backward compatibility.