BUG: Loading co-ordinates larger than 32768
Data for the left/top/right/bottom boundaries of the ROI are stored as
either int16 or uint16, with negative co-ordinates and values larger than
2^15 mapping to the same values.

We now make a best guess as to whether the value is negative or very big
by checking whether it exceeds the threshold 65036. Values below this remain
large positive numbers. Values greater than this become negative values
no more negative than -500.

Closes #13.